+++ build.rs
@@ -69,6 +69,7 @@ fn main() {
     opencv_config
         .cpp(true)
         .file("native/opencv-wrapper.cc")
+        .file("native/utils.cc")
         .include("native")
         .include(opencv_include());
