+++ native/utils.h
@@ -3,30 +3,11 @@
 
 #include <vector>
 #include <opencv2/core.hpp>
-#include "opencv-wrapper.h"
+#include "common.h"
 
-// =============================================================================
-//   Utils
-// =============================================================================
-
-void cv_to_ffi(const cv::Rect& source, Rect* dest){
-    dest->x = source.x;
-    dest->y = source.y;
-    dest->width = source.width;
-    dest->height = source.height;
-}
-
-void cv_to_ffi(const cv::Point& source, Point2i* dest){
-    dest->x = source.x;
-    dest->y = source.y;
-};
-
-void cv_to_ffi(const std::vector<double>& source, CVec<double>* dest) {
-    size_t num = source.size();
-    dest->size = num;
-    dest->array = (double*) malloc(num * sizeof(double));
-    ::memcpy(dest->array, source.data(), num * sizeof(double));
-}
+void cv_to_ffi(const cv::Rect& source, Rect* dest);
+void cv_to_ffi(const cv::Point& source, Point2i* dest);
+void cv_to_ffi(const std::vector<double>& source, CVec<double>* dest);
 
 template <typename T, typename U>
 void cv_to_ffi(const std::vector<T>& source, CVec<U>* dest)
